move v0 http client configuration to dedicated package
